Refactor walkthrough helper into input.go, add tests for format.go
